+++ resources/views/storefront/partials/image-banner.blade.php
@@ -101,7 +101,7 @@
                             <source media="(max-width: 639px)" srcset="{{ \App\Support\CompanyMedia::publicUrl($slide->image_mobile, $company) }}">
                         @endif
                         <img
-                            class="block h-full w-full object-cover"
+                            class="block h-full w-full"
                             src="{{ \App\Support\CompanyMedia::publicUrl($slide->image, $company) }}"
                             alt="{{ $slideHref ? $company->name.' promotion' : '' }}"
                             width="1920"